Added setting to be able to scale the resolution of OpenXR views (#44)
Signed-off-by: moraaar <moraaar@amazon.com>

The setting can make use of platform prefixes to only target one platform, for example:

````
{
    "O3DE": {
        "Atom": {
            "OpenXR": {
                "android_ViewResolutionScale": 0.5
            }
        }
    }
}
````

NOTE: This setting is a temporary measure that will help to control the performance when running on Quest 2 by lowering the resolution. Further optimizations to make the VR pipeline more performant overall will be done in the future.