Updated the Link query logic in the RobotImporter
Joints in which a link is parent can be queried using the new `GetJointsForParentLink` function.

Joints in which a link is child can be queried using the new `GetJointsForChildLink` function.

A VisitJointsForModel and VisitLinksForModel function has been added to allow visting each joint or link on model.
It supports visiting joints or links on nested models as well

Fixed the URDF Parser Test to account for link and joints changes when using libsdformat for parsing a URDF.

SDFormat URDF parser uses joint reduction to combine two links joined together by fixed joint into the parent link itself.

That operation results in the removal of the child link and the joint.
SDFormat creates a Frame for the removed child link and joint using their "name" fields, so it is possible to still discover information about the child link.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>