Addressed feedback for URDF Import using SDFormat PR
Added a "/O3DE/ROS2/SdfAssetBuilder/URDFPreserveFixedJoint" setting which is used to preserve fixed joints when parsing URDFs

In order to use the URDFPreserveFixedJoint setting, the UrdfParser `Parse*` functions have been updated to accept an sdf::ParserConfig which can be used to set the preserve fixed joint setting.

Refactored the RobotImporter.cpp ResolveURDFPath function to support "model://" URI prefix when importing through the Widget.

Fixed the GetAllLinks function population of the SDF LinkMap to not use a dangling reference to the Link Name.

Fixed building of ROS2 Gem when Unity builds are turned off.
The SdfAssetBuilderName was being used in the SdfAssetBuilderSettings.cpp but it was only declared in the SdfAssetBuilder.cpp file

Moved the function that converts a sdf::Error vector into an AZStd::string to a Common ErrorUtils.cpp/h file.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>