Updated the `CreatePrefabTemplateFromURDF` link hierarchy logic
As an SDF file represents a graph structure, there is not necessarily a single root link within it.

Therefore the `CreatePrefabTemplateFromURDF` has been updated to gather a set of all links that aren't parented within an SDF file and then use the first link as the entity for the Robot Control component and for setting the spawn position.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>