Updated the Joint Visitor for SDF joints
Joints which do not have a valid parent or child link will be skipped during visting using the `ROS2::Utils::VisitJoints` API.

Updated the URDF Prefab Makaer to use the VisitJoints API to only visit joints which have links that are part of the SDF.

With this change, URDFs that are parsed with a root link that has the name of "world" will now skip over visting joints where the parent link being referenced is the "world" link.

This is needed because the libsdformat URDF parser skips converstion of root links with the name of "world" when converting to SDF, therefore the joints which references the "world" link also needs to be skipped.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>