Updated the URDFPrefabMaker to avoid referencing SDF link names (#523)
Updated the URDFPrefabMaker to avoid referencing SDF link names as they are not globally unique

The `CreatePrefabTemplateFromUrdfOrSdf` function has been modified to create a JointMapper object which stores a mapping of joints to all their parent and child links.
It also creates a mapping of joints to the model they are directly attached to.

A similar LinkMapper structure has been added for SDF links, which stores a mapping of the link to their attached model.

Instead of `CreatePrefabTemplateFromUrdfOrSdf` creating a mapping of SDF link name to O3DE Entity, that mapping has now been updated toi map the SDF link pointer to O3DE Entity.
This ensures that SDF links with same name are not lost when mapping to O3DE entities.
This is needed as SDF links and joints are only required to be unique relative to their attached model.
Two links on different models, can have the same link name (http://sdformat.org/tutorials?tut=composition_proposal#1-6-proposed-parsing-stages)

Updated the Robot Importer Visitor callback to require "ModelStack" parameters, which passes in the stack of `sdf::Model` objects that were visited on the way to the sdf element being visited be that a link, joint or nested model.
Also updated the GetAllJoints and GetAllLinks function to store a fully qualified name for the joint/link including the ancestor model names when returning the name to joint and name to link map respectively.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

Addressed PR feedback for o3de/o3de-extra#523 around invalid text and grammar

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>