SDF Nested Model Support Changes (#561)
Updated the Check Asset Page to not add Assets with empty source paths or null Asset IDs to the list of resolved assets

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

Added a `GetModelContainingModel` function that can return the parent model of a nested model

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

Updated the URDFPrefabMaker to create an Entity for which represents the URDF/SDF model itself.

The AddEntitiesForLink function has been updated to to pass in the model that the link is attached to and in order to allow the CollidersMaker, ArticulationsMaker and SensorsMaker have access to the Model

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

Addressed sentence typos in comments.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

Fixed the joint iteration for loop in `CreatePrefabTemplateFromUrdfOrSdf`

The code in hee for loop was converted from a lambda to a for loop, but the return statements weren't converted.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>