Fixed parenting of SDF model and link entity transforms
The pose information from the SDF is a value relative to it's attached frame.
The `GetLocalTransfromURDF` function was misnamed and returning the a local transform for any SDF link supplied to the function.

Also the `PrefabMakerUtils::SetEntityParent` function which is used to attach a child entity to a parent entity, preserves the world transform of the child entity. Therefore any pose information on child entities such as nested models or links on a model would be set in world space and not in local space.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>