Fixed the setup of the parent child entity hierarchy between SDF links
The parent child hierarchy is established through an SDF joint element which references a child and a parent link.

The hieararchy itself shouldn't be used to adjust the world transform between the two links, only to have the child link entity to be under parent link entity for organization purposes.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>