New OpenXR ActionSets and InteractionProfiles assets
Per this RFC:
https://github.com/o3de/sig-graphics-audio/blob/main/rfcs/OpenXRActionsAndSpaces/rfc_openxr_actions_and_spaces.md
This commit is the first of several more that will be commited in different batches for the sake
of making it easier for code reviewers.

This commit has no runtime implications, other that registering
the OpenXRInteractionProfilesAsset and the OpenXRActionSetsAsset,
 along with the asset builder named OpenXRAssetsBuilder, which is in charge
 of validating and building both asset types.

OpenXRInteractionProfilesAsset will have the `*.xrprofiles` file extension.
An example asset is provided as the following path:
`Gems/OpenXRVk/Assets/OpenXRVk/system.xrprofiles`

OpenXRActionSetsAsset will have the `*.rxactions` file extension.
An example asset is provided as the following path:
`Gems/OpenXRVk/Assets/OpenXRVk/default.xractions`
This asset has a job dependency on `Gems/OpenXRVk/Assets/OpenXRVk/system.xrprofiles`.

Both asset types are also editable with the `Asset Editor`.

The Assets mentioned above will be compiled and placed in the asset cache
but they are unused at the moment for the sake of an easier code review.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>