Upgraded openxr dependency to 1.1.41.
This update makes it straightforward to build O3DE
applications for Meta Quest devices, because the developer
is not required anymore to download and decompress the `Oculus OpenXR Mobile SDK`.

To understand this change, let's review a little bit of history:
When making O3DE apps for Meta Quest devices, the developer
was instructed to manually download and decompress the
 `Oculus OpenXR Mobile SDK` into a particular directory of the `OpenXRVk` Gem.
Those instructions worked well until SDK v62.

Starting SDK v63+ (It's currently at v72)
Meta stopped distributing pre-compiled OpenXR Loader libraries, and now applications are simply
required to get the OpenXR Loader directly from github:
https://github.com/KhronosGroup/OpenXR-SDK

In particular, this PR picks version 1.1.41 because that's the version used
by `v72` of `Oculus OpenXR Mobile SDK`.
In other words, with this change, OpenXR is now supported in its purest form,
because there are no custom steps for a particular vendor.

Updated to rev2 of openxr-1.1.41 package
which was recompiled by the 3p-package-source
with NDK 25.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>